# apigen/openapi/bucket.openapi.yaml
openapi: 3.0.3
info:
  title: Bucket Credential API (DRS Server)
  version: 1.0.0
  description: |
    Bucket credential and scope management API used by drs-server compatible clients.
servers:
  - url: /
tags:
  - name: bucket
paths:
  /data/buckets:
    get:
      tags:
        - bucket
      summary: List configured bucket credentials and org/project assignments
      operationId: listBuckets
      responses:
        '200':
          description: Bucket map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BucketsResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
    put:
      tags:
        - bucket
      summary: Upsert bucket credential and claim org/project scope
      operationId: putBucket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutBucketRequest'
      responses:
        '201':
          description: Bucket credential and scope created/updated
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '409':
          description: Scope already assigned to a different bucket/path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
  /data/buckets/{bucket}:
    delete:
      tags:
        - bucket
      summary: Delete bucket credential
      operationId: deleteBucket
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
  /data/buckets/{bucket}/scopes:
    get:
      tags:
        - bucket
      summary: List org/project scopes for a bucket
      operationId: listBucketScopes
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bucket scopes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BucketScopeResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '404':
          description: Bucket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
    post:
      tags:
        - bucket
      summary: Add an org/project scope to an existing bucket
      operationId: addBucketScope
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBucketScopeRequest'
      responses:
        '201':
          description: Scope created
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '404':
          description: Bucket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '409':
          description: Scope already assigned to a different bucket/path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
    delete:
      tags:
        - bucket
      summary: Remove a specific scope from an existing bucket
      operationId: deleteBucketScope
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
        - name: organization
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          allowEmptyValue: true
          schema:
            type: string
        - name: project_id
          in: query
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Scope removed
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '404':
          description: Scope not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
  /data/projects/{organization}/{project_id}:
    delete:
      tags:
        - bucket
      summary: Delete Syfon records and bucket scopes for a project
      operationId: deleteProjectData
      parameters:
        - name: organization
          in: path
          required: true
          schema:
            type: string
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project Syfon records and bucket scopes removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteProjectDataResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
components:
  schemas:
    APIError:
      $ref: './error.openapi.yaml#/components/schemas/APIError'
    PutBucketRequest:
      type: object
      required:
        - bucket
        - organization
        - project_id
      properties:
        bucket:
          type: string
        provider:
          type: string
          description: Storage backend provider (s3|gcs|azure|file). Defaults to s3.
        region:
          type: string
        access_key:
          type: string
        secret_key:
          type: string
        endpoint:
          type: string
        organization:
          type: string
        project_id:
          type: string
        path:
          type: string
          description: Optional s3://bucket/prefix path override
    AddBucketScopeRequest:
      type: object
      required:
        - organization
        - project_id
      properties:
        organization:
          type: string
        project_id:
          type: string
        path:
          type: string
          description: Optional s3://bucket/prefix path override
    BucketMetadata:
      type: object
      properties:
        bucket:
          type: string
        endpoint_url:
          type: string
        provider:
          type: string
        region:
          type: string
        programs:
          type: array
          items:
            type: string
    BucketsResponse:
      type: object
      required:
        - S3_BUCKETS
      properties:
        S3_BUCKETS:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/BucketMetadata'
    DeleteProjectDataResponse:
      type: object
      required:
        - organization
        - project_id
        - deleted_objects
        - deleted_bucket_scopes
      properties:
        organization:
          type: string
        project_id:
          type: string
        deleted_objects:
          type: integer
        deleted_bucket_scopes:
          type: integer
    BucketScopeResponse:
      type: object
      required:
        - organization
        - project_id
      properties:
        organization:
          type: string
        project_id:
          type: string
        path:
          type: string
          description: Scope path prefix